font-patcher: Add option to download symbol glyphs
[why]
People often just download the font-patcher script and then the patching
fails because the symbol glyphs are missing.

A typical comment is that a complete git clone is needed in that case,
which is rediculous bandwidth. Just downloading the needed files seems
rather complicated.

[how]
Add option --glyphdownload that (tries to) download just the needed
glyph files. This is just bare basic stuff.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>